// src/internal/az/azreq/grpc/v1/generated/proto/v1/pdp.ts
export interface PolicyStore {
    /**
     * @generated from protobuf field: string Kind = 1 [json_name = "Kind"];
     */
    kind: string;
    /**
     * @generated from protobuf field: string ID = 2 [json_name = "ID"];
     */
    iD: string;
}
/**
 * Principal represents the entity making the request.
 *
 * @generated from protobuf message policydecisionpoint.Principal
 */
export interface Principal {
    /**
     * @generated from protobuf field: string Type = 1 [json_name = "Type"];
     */
    type: string;
    /**
     * @generated from protobuf field: string ID = 2 [json_name = "ID"];
     */
    iD: string;
    /**
     * @generated from protobuf field: optional string Source = 3 [json_name = "Source"];
     */
    source?: string;
    /**
     * @generated from protobuf field: optional string IdentityToken = 4 [json_name = "IdentityToken"];
     */
    identityToken?: string;
    /**
     * @generated from protobuf field: optional string AccessToken = 5 [json_name = "AccessToken"];
     */
    accessToken?: string;
}
/**
 * Entities represent the entities provided in the context for the authorization decision.
 *
 * @generated from protobuf message policydecisionpoint.Entities
 */
export interface Entities {
    /**
     * @generated from protobuf field: string Schema = 1 [json_name = "Schema"];
     */
    schema: string;
    /**
     * @generated from protobuf field: repeated google.protobuf.Struct Items = 2 [json_name = "Items"];
     */
    items: Struct[];
}
/**
 * Subject is the entity on which the authorization decision is made.
 *
 * @generated from protobuf message policydecisionpoint.Subject
 */
export interface Subject {
    /**
     * @generated from protobuf field: string Type = 1 [json_name = "Type"];
     */
    type: string;
    /**
     * @generated from protobuf field: string ID = 2 [json_name = "ID"];
     */
    iD: string;
    /**
     * @generated from protobuf field: optional string Source = 3 [json_name = "Source"];
     */
    source?: string;
    /**
     * @generated from protobuf field: optional google.protobuf.Struct Properties = 4 [json_name = "Properties"];
     */
    properties?: Struct;
}
/**
 * Resource is the entity on which the authorization decision is made.
 *
 * @generated from protobuf message policydecisionpoint.Resource
 */
export interface Resource {
    /**
     * @generated from protobuf field: string Type = 1 [json_name = "Type"];
     */
    type: string;
    /**
     * @generated from protobuf field: string ID = 2 [json_name = "ID"];
     */
    iD: string;
    /**
     * @generated from protobuf field: optional google.protobuf.Struct Properties = 3 [json_name = "Properties"];
     */
    properties?: Struct;
}
/**
 * Action is the operation on which the authorization decision is made.
 *
 * @generated from protobuf message policydecisionpoint.Action
 */
export interface Action {
    /**
     * @generated from protobuf field: string Name = 1 [json_name = "Name"];
     */
    name: string;
    /**
     * @generated from protobuf field: optional google.protobuf.Struct Properties = 2 [json_name = "Properties"];
     */
    properties?: Struct;
}
// AuthorizationCheck Request

/**
 * AuthorizationModelRequest is the input context for making the authorization decision.
 *
 * @generated from protobuf message policydecisionpoint.AuthorizationModelRequest
 */
export interface AuthorizationModelRequest {
    /**
     * @generated from protobuf field: int64 ZoneID = 1 [json_name = "ZoneID"];
     */
    zoneID: bigint;
    /**
     * @generated from protobuf field: policydecisionpoint.PolicyStore PolicyStore = 2 [json_name = "PolicyStore"];
     */
    policyStore?: PolicyStore;
    /**
     * @generated from protobuf field: optional policydecisionpoint.Principal Principal = 3 [json_name = "Principal"];
     */
    principal?: Principal;
    /**
     * @generated from protobuf field: optional policydecisionpoint.Entities Entities = 4 [json_name = "Entities"];
     */
    entities?: Entities;
}
/**
 * EvaluationRequest represents the request to evaluate the authorization decision.
 *
 * @generated from protobuf message policydecisionpoint.EvaluationRequest
 */
export interface EvaluationRequest {
    /**
     * @generated from protobuf field: optional string RequestID = 1 [json_name = "RequestID"];
     */
    requestID?: string;
    /**
     * @generated from protobuf field: optional policydecisionpoint.Subject Subject = 2 [json_name = "Subject"];
     */
    subject?: Subject;
    /**
     * @generated from protobuf field: optional policydecisionpoint.Resource Resource = 3 [json_name = "Resource"];
     */
    resource?: Resource;
    /**
     * @generated from protobuf field: optional policydecisionpoint.Action Action = 4 [json_name = "Action"];
     */
    action?: Action;
    /**
     * @generated from protobuf field: optional google.protobuf.Struct Context = 5 [json_name = "Context"];
     */
    context?: Struct;
}
/**
 * AuthorizationCheckRequest represents the request to perform an authorization decision.
 *
 * @generated from protobuf message policydecisionpoint.AuthorizationCheckRequest
 */
export interface AuthorizationCheckRequest {
    /**
     * @generated from protobuf field: policydecisionpoint.AuthorizationModelRequest AuthorizationModel = 1 [json_name = "AuthorizationModel"];
     */
    authorizationModel?: AuthorizationModelRequest;
    /**
     * @generated from protobuf field: optional string RequestID = 2 [json_name = "RequestID"];
     */
    requestID?: string;
    /**
     * @generated from protobuf field: optional policydecisionpoint.Subject Subject = 3 [json_name = "Subject"];
     */
    subject?: Subject;
    /**
     * @generated from protobuf field: optional policydecisionpoint.Resource Resource = 4 [json_name = "Resource"];
     */
    resource?: Resource;
    /**
     * @generated from protobuf field: optional policydecisionpoint.Action Action = 5 [json_name = "Action"];
     */
    action?: Action;
    /**
     * @generated from protobuf field: optional google.protobuf.Struct Context = 6 [json_name = "Context"];
     */
    context?: Struct;
    /**
     * @generated from protobuf field: repeated policydecisionpoint.EvaluationRequest Evaluations = 7 [json_name = "Evaluations"];
     */
    evaluations: EvaluationRequest[];
}
// AuthorizationCheck Response

/**
 * ReasonResponse provides the rationale for the response.
 *
 * @generated from protobuf message policydecisionpoint.ReasonResponse
 */
export interface ReasonResponse {
    /**
     * @generated from protobuf field: string Code = 1 [json_name = "Code"];
     */
    code: string;
    /**
     * @generated from protobuf field: string Message = 2 [json_name = "Message"];
     */
    message: string;
}
/**
 * ContextResponse represents the context included in the response.
 *
 * @generated from protobuf message policydecisionpoint.ContextResponse
 */
export interface ContextResponse {
    /**
     * @generated from protobuf field: string ID = 1 [json_name = "ID"];
     */
    iD: string;
    /**
     * @generated from protobuf field: policydecisionpoint.ReasonResponse ReasonAdmin = 2 [json_name = "ReasonAdmin"];
     */
    reasonAdmin?: ReasonResponse;
    /**
     * @generated from protobuf field: policydecisionpoint.ReasonResponse ReasonUser = 3 [json_name = "ReasonUser"];
     */
    reasonUser?: ReasonResponse;
}
/**
 * EvaluationResponse represents the result of the evaluation process.
 *
 * @generated from protobuf message policydecisionpoint.EvaluationResponse
 */
export interface EvaluationResponse {
    /**
     * @generated from protobuf field: bool Decision = 1 [json_name = "Decision"];
     */
    decision: boolean;
    /**
     * @generated from protobuf field: optional string RequestID = 2 [json_name = "RequestID"];
     */
    requestID?: string;
    /**
     * @generated from protobuf field: optional policydecisionpoint.ContextResponse Context = 3 [json_name = "Context"];
     */
    context?: ContextResponse;
}
/**
 * AuthorizationCheckResponse represents the outcome of the authorization decision.
 *
 * @generated from protobuf message policydecisionpoint.AuthorizationCheckResponse
 */
export interface AuthorizationCheckResponse {
    /**
     * @generated from protobuf field: bool Decision = 1 [json_name = "Decision"];
     */
    decision: boolean;
    /**
     * @generated from protobuf field: optional string RequestID = 2 [json_name = "RequestID"];
     */
    requestID?: string;
    /**
     * @generated from protobuf field: optional policydecisionpoint.ContextResponse Context = 3 [json_name = "Context"];
     */
    context?: ContextResponse;
    /**
     * @generated from protobuf field: repeated policydecisionpoint.EvaluationResponse Evaluations = 4 [json_name = "Evaluations"];
     */
    evaluations: EvaluationResponse[];
}
// @generated message type with reflection information, may provide speed optimized methods
class PolicyStore$Type extends MessageType<PolicyStore> {
    constructor() {
        super("policydecisionpoint.PolicyStore", [
            { no: 1, name: "Kind", kind: "scalar", jsonName: "Kind", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "ID", kind: "scalar", jsonName: "ID", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<PolicyStore>): PolicyStore {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.kind = "";
        message.iD = "";
        if (value !== undefined)
            reflectionMergePartial<PolicyStore>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PolicyStore): PolicyStore {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string Kind = 1 [json_name = "Kind"];*/ 1:
                    message.kind = reader.string();
                    break;
                case /* string ID = 2 [json_name = "ID"];*/ 2:
                    message.iD = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PolicyStore, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string Kind = 1 [json_name = "Kind"]; */
        if (message.kind !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.kind);
        /* string ID = 2 [json_name = "ID"]; */
        if (message.iD !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.iD);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message policydecisionpoint.PolicyStore
 */
export const PolicyStore = new PolicyStore$Type();
// ...
export const Principal = new Principal$Type();
// ...
export const Entities = new Entities$Type();
// ...
export const Subject = new Subject$Type();
// ...
export const Resource = new Resource$Type();
// ...
export const Action = new Action$Type();
// ...
export const AuthorizationModelRequest = new AuthorizationModelRequest$Type();
// ...
export const EvaluationRequest = new EvaluationRequest$Type();
// ...
export const AuthorizationCheckRequest = new AuthorizationCheckRequest$Type();
// ...
export const ReasonResponse = new ReasonResponse$Type();
// ...
export const ContextResponse = new ContextResponse$Type();
// ...
export const EvaluationResponse = new EvaluationResponse$Type();
// ...
export const AuthorizationCheckResponse = new AuthorizationCheckResponse$Type();